syncbridge-backend/first version/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.52
alembic==1.9.1
pydantic==1.10.7
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
python-jose==3.3.0
aiofiles==23.1.0
